renaming Error to ErrorComp
