+++ components/ErrorComp.tsx
@@ -4,7 +4,7 @@ type Props = {
   errorText: string;
 };
 
-const Error: React.FC<Props> = ({ errorText }) => {
+const ErrorComp: React.FC<Props> = ({ errorText }) => {
   return (
     <View className="h-full bg-white items-center pt-64 px-3">
       <Text className="text-lg font-semibold mb-2">An error occured:</Text>
@@ -13,4 +13,4 @@ const Error: React.FC<Props> = ({ errorText }) => {
   );
 };
 
-export default Error;
+export default ErrorComp;
